Fixed issue with blanks in job schedules
There was an issue with the start_at and last_at options
causing a problem with the Rufus scheduler when they
were blank instead of nil.  It looks like the factories
were creating them as nil but the forms in the app were
creating them as blank.  I would prefer everything
being saved as nil so I added the nilify_blanks gem to
all of the models.