BugFix - Duplicate space was not honoring checkboxes
Two issues were at work here.  The first was that false.blank? evaluates to true, so the way that
the defaults were being set in models/tpl_birst_duplicate_space would override the user if they
unchecked a box.  The second more difficult problem to solve was that each time the job template
was being updated via the job_spec view, it was creating a new job template id.  Fixing it required
adding the job_template_id as a hidden form element and finagling the params creation
in the controller.