BugFix - New LaunchJob id
There was an issue where the LaunchedJob id was being reused each time the
Scheduler ran the job.  This wasn't detected in initial tests because they would
start and stop the scheduler between Rufus runs.  This fix creates a new JobHandler
in the LaunchedJob namespace that creates a new LaunchedJob instance each
time Rufus executes the job.